editor/libeditor/html/tests/test_bug597784.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=597784
-->
<head>
  <title>Test for Bug 597784</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=597784">Mozilla Bug 597784</a>
<p id="display"></p>
<div id="content"></div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 597784 **/

SimpleTest.waitForExplicitFinish();
addLoadEvent(function() {
  document.designMode = "on";
  var content = document.getElementById("content");
  getSelection().collapse(content, 0);
  var html1 = "<test:tag>test:tag</test:tag>";
  var html2 = "<a href=\"http://mozilla.org/\" test:attr=\"test:attr\" custom=\"value\">link</a>";
  document.execCommand("insertHTML", false, html1);
  document.execCommand("insertHTML", false, html2);
  is(content.innerHTML, html1 + html2,
     "The custom tags and attributes should be inserted into the document using the insertHTML command");
  SimpleTest.finish();
});

</script>
</pre>
</body>
</html>
